+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/table_part/extension_list/extension.rb
@@ -1,10 +1,13 @@
 # frozen_string_literal: true
 
+require_relative 'extension/data_validations'
 require_relative 'extension/sparkline_groups'
 require_relative 'extension/x14_table'
 module OoxmlParser
   # Class for `ext` data
   class Extension < OOXMLDocumentObject
+    # @return [DataValidations] list of data validations
+    attr_accessor :data_validations
     # @return [X14Table] table data in x14 namespace
     attr_accessor :table
     # @return [SparklineGroups] list of groups
@@ -16,6 +19,8 @@ module OoxmlParser
     def parse(node)
       node.xpath('*').each do |column_node|
         case column_node.name
+        when 'dataValidations'
+          @data_validations = DataValidations.new(parent: self).parse(column_node)
         when 'table'
           @table = X14Table.new(parent: self).parse(column_node)
         when 'sparklineGroups'